Teams with leagues

+++ src/main/java/com/example/thymeleaf/model/entity/League.java
@@ -16,13 +16,9 @@ import javax.persistence.Table;
 public class League {
 
     @Id
-    private Long id;
-
+    private Integer id;
     private String name;
-
     private String country;
-
     private String code;
-
     private Boolean isActive;
 }
